refactor questions editor into a seperate component
